Make clock_t unsigned int everywhere. Ok: matt@, mrg@

+++ sys/arch/hppa/include/ansi.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: ansi.h,v 1.13 2011/07/17 20:54:41 joerg Exp $	*/
+/*	$NetBSD: ansi.h,v 1.14 2015/06/17 14:32:31 martin Exp $	*/
 
 /*	$OpenBSD: ansi.h,v 1.4 2000/02/22 17:29:12 millert Exp $	*/
 
@@ -48,7 +48,7 @@
  *	#undef	_BSD_SIZE_T_
  *	#endif
  */
-#define	_BSD_CLOCK_T_	unsigned long int	/* clock() */
+#define	_BSD_CLOCK_T_	unsigned int		/* clock() */
 #define	_BSD_PTRDIFF_T_	long int		/* ptr1 - ptr2 */
 #define	_BSD_SIZE_T_	unsigned long int	/* sizeof() */
 #define	_BSD_SSIZE_T_	long int		/* byte count or error */
